fix: Resolve PHP coding standard violations
- Remove blank line at end of control structure in CleanupDemoDataCommand
- Split ProcessOpenBankingDepositActivity into separate file (PSR-4 compliance)
- Each class must be in its own file per PSR standards
- Fix missing newline at end of file
